Fix - missing structattrs for CorpusActionModel,...
... also remove unused attribute on the client-side
